fix plugin naming
