+++ packages/nextjs-mf/src/plugins/ChildFederationPlugin.ts
@@ -125,12 +125,12 @@ export class ChildFederationPlugin {
         ];
       } else if (compiler.options.name === 'server') {
         const {
-          NodeFederationPlugin,
           StreamingTargetPlugin,
+          NodeFederationPlugin,
         } = require('@module-federation/node');
 
         plugins = [
-          new StreamingTargetPlugin(federationPluginOptions, {
+          new NodeFederationPlugin(federationPluginOptions, {
             ModuleFederationPlugin: FederationPlugin,
           }),
           new webpack.node.NodeTemplatePlugin(childOutput),
@@ -144,7 +144,7 @@ export class ChildFederationPlugin {
             'react/jsx-dev-runtime',
           ]),
           // new LoaderTargetPlugin('async-node'),
-          new NodeFederationPlugin(federationPluginOptions, {
+          new StreamingTargetPlugin(federationPluginOptions, {
             ModuleFederationPlugin: webpack.container.ModuleFederationPlugin,
           }),
           new LibraryPlugin(federationPluginOptions.library?.type as string),
