+++ packages/node/src/index.ts
@@ -1,2 +1,2 @@
-export { default as NodeFederationPlugin } from './plugins/NodeFederationPlugin';
 export { default as StreamingTargetPlugin } from './plugins/StreamingTargetPlugin';
+export { default as NodeFederationPlugin } from './plugins/NodeFederationPlugin';
